New Calculator Program is under UpdatedCalculator.java, uses CalcMethods.java

+++ src/stdio/CalcMethods.java
@@ -2,16 +2,16 @@ package stdio;
 
 class CalcMethods {
 	
-	int add(int a, int b) {
+	double add(double a, double b) {
 		return a+b;
 	}
-	int subtract(int a, int b) {
+	double subtract(double a, double b) {
 		return a-b;
 	}
-	int multiply(int a, int b) {
+	double multiply(double a, double b) {
 		return a*b;
 	}	
-	int divide(int a, int b) {
+	double divide(double a, double b) {
 		return a/b;
 	}
